backend/tests/conftest.py
import pytest
from fastapi.testclient import TestClient
from sqlalchemy.pool import StaticPool
from sqlmodel import Session, SQLModel, create_engine

from app.db.session import get_db
from app.main import app

VALID_TRIP_REQUEST = {
    "departure": "JFK",
    "destination": "CDG",
    "pto_days_remaining": 5,
    "max_flight_budget": 1000.0,
    "company_holidays": [],
}


@pytest.fixture
def client():
    engine = create_engine(
        "sqlite://",
        connect_args={"check_same_thread": False},
        poolclass=StaticPool,
    )
    SQLModel.metadata.create_all(engine)

    def override_get_db():
        with Session(engine) as session:
            yield session

    app.dependency_overrides[get_db] = override_get_db
    yield TestClient(app)
    app.dependency_overrides.clear()
