backend/requirements.txt
fastapi[standard]
black
ruff
pre-commit
pydantic
pydantic-settings
google-search-results
pytest
pytest-asyncio
respx
langchain
langchain-core
langchain-openai
langgraph
sqlmodel
psycopg2-binary
alembic
python-jose[cryptography]
bcrypt